requirements.txt
prefect
prefect-aws
httpx
GitPython
pymongo
psutil
setuptools
dotenv
wheel
plotly
kaleido
numpy
requests
pydantic-ai
logfire[pymongo,httpx]
click
demjson3
tiktoken
tinydb
aiotinydb
BetterJSONStorage
uvicorn